Remove bigkicker to make room for solenoid
For now, the solenoid is being added by manually tweaking the .lat file output
from MADX. If the bigkicker marker is not removed, the PyORBIT will make the
ring longer to make room for scbdsol.

In the future, we need to modify the original MADX script to include the solenoid.
It should have a finite length; PyORBIT does not have an effective length
parameter like MADX (KSI). It can probably just be set to zero in MADX to
set the x/y tunes.