2023-03-10 SNS-BTF RFQ-HZ34a
A measured 5D+2D bunch (2022-06-26) was backtracked from HZ04 to RFQ; now we forward track from
RFQ to HZ34a. This gives the same results as tracking from HZ04 to HZ34a; however,
this allows us to directly compare the first meter of transport with the previous
simulation, which decorrelated the initial bunch at the RFQ.

The beam current was set to the measured value of 25.5 mA.

This can serve as a full benchmark.