Check synchronous particle time if not starting at origin
To be safe, PyORBIT bunches should be saved with the header
lines included; this will load the synchronous particle time
when bunch.readBunch() is called. The new method will check
the design synchronous particle time and compare with the
bunch; if they are different, there is the option to reset
the time to the design value.